# admin/api/openapi.yaml
openapi: 3.0.2
info:
  contact:
    url: https://github.com/SecurityPPH/watchman
  description: |
    Watchman is an HTTP API and Go library to download, parse and offer search functions over numerous trade sanction lists from the United States, European Union governments, agencies, and non profits for complying with regional laws. Also included is a web UI and async webhook notification service to initiate processes on remote systems.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Watchman Admin API
  version: v1
servers:
- description: Local development
  url: http://localhost:9094
tags:
- description: Watchman endpoints which are only meant to be exposed for admin dashboards
    and operations.
  name: Admin
paths:
  /version:
    get:
      description: Show the current version of Watchman
      operationId: getVersion
      responses:
        "200":
          content:
            text/plain:
              schema:
                example: v0.13.1
                type: string
          description: The current version running
      summary: Get Version
      tags:
      - Admin
  /data/refresh:
    post:
      operationId: refreshData
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataRefresh'
          description: Data successfully refreshed
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: See error message
      summary: Download and reindex all data sources
      tags:
      - Admin
  /debug/sdn/{sdnId}:
    get:
      description: Get an SDN and search index debug information
      operationId: debugSDN
      parameters:
      - description: SDN ID
        explode: false
        in: path
        name: sdnId
        required: true
        schema:
          example: 564dd7d1
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebugSDN'
          description: SDN with debug information
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: See error message
      summary: Debug SDN
      tags:
      - Admin
components:
  schemas:
    SDNDebugMetadata:
      example:
        indexedName: mohammad moghisseh
        parsedRemarksId: "3603251708570001"
      properties:
        indexedName:
          description: Exact text stored in our index used for string ranking
          example: mohammad moghisseh
          type: string
        parsedRemarksId:
          description: ID parsed from remarks field
          example: "3603251708570001"
          type: string
    DebugSDN:
      example:
        debug:
          indexedName: mohammad moghisseh
          parsedRemarksId: "3603251708570001"
        SDN:
          match: 0.91
          entityID: "1231"
          programs:
          - CUBA
          sdnName: BANCO NACIONAL DE CUBA
          title: Title of an individual
          remarks: remarks
      properties:
        SDN:
          $ref: '#/components/schemas/OfacSDN'
        debug:
          $ref: '#/components/schemas/SDNDebugMetadata'
    DataRefresh:
      example:
        SDNs: 13131
        altNames: 322
        addresses: 4155
        deniedPersons: 5889
        bisEntities: 6831
        sectoralSanctions: 667
        timestamp: 2000-01-23T04:56:07.000+00:00
      properties:
        SDNs:
          description: Count of OFAC SDNs after index
          example: 13131
          type: integer
        altNames:
          description: Count of OFAC alternate names after index
          example: 322
          type: integer
        addresses:
          description: Count of OFAC Addresses after index
          example: 4155
          type: integer
        sectoralSanctions:
          description: Count of SSI entities after index
          example: 667
          type: integer
        deniedPersons:
          description: Count of BSL denied persons after index
          example: 5889
          type: integer
        bisEntities:
          description: Count of BIS entities after index
          example: 6831
          type: integer
        timestamp:
          format: date-time
          type: string
    Error:
      properties:
        error:
          description: An error message describing the problem intended for humans.
          example: Example error, see description
          type: string
      required:
      - error
    OfacSDN:
      description: Specially designated national from OFAC list
      example:
        match: 0.91
        entityID: "1231"
        programs:
        - CUBA
        sdnName: BANCO NACIONAL DE CUBA
        title: Title of an individual
        remarks: remarks
      properties:
        entityID:
          example: "1231"
          type: string
        sdnName:
          example: BANCO NACIONAL DE CUBA
          type: string
        sdnType:
          $ref: '#/components/schemas/SdnType'
        programs:
          description: Programs is the sanction programs this SDN was added from
          example:
          - CUBA
          items:
            type: string
          type: array
        title:
          example: Title of an individual
          type: string
        remarks:
          type: string
        match:
          description: Remarks on SDN and often additional information about the SDN
          example: 0.91
          type: number
    SdnType:
      description: Used for classifying SDNs — typically represents an individual
        or company
      enum:
      - individual
      - entity
      - vessel
      - aircraft
      type: string